Solution to Input Two Floats and Print Their Product

+++ Input Two Floats and Print Their Product@@6788d5012db25efec946f93e/code.c
@@ -5,8 +5,8 @@ char* welcome() {
 }
 
 int main() {
-    int a,b;
+    float a,b;
     scanf("%f %f", &a, &b);
-    printf("Product: %f",a+b);
+    printf("Product: %.2f",a+b);
     return 0;
 }
